Move pe to separate module
