+++ src/modules/positional_encodings.py
@@ -1,0 +1,36 @@
+import math
+
+import torch
+import torch.nn as nn
+
+
+class SinusoidalPositionalEncoding(nn.Module):
+    def __init__(self, 
+                 d_model: int, 
+                 max_len: int, 
+                 device, 
+                 dropout: float = 0.0) -> None:
+        super().__init__()
+        self.dropout = nn.Dropout(p=dropout)
+
+        position = torch.arange(max_len).unsqueeze(1)
+        # ```d_model // 2``` elements
+        div_term = torch.exp(torch.arange(0, d_model, 2) * (-math.log(10000.0) / d_model))
+        pe = torch.zeros(max_len, 1, d_model)
+        pe[:, 0, 0::2] = torch.sin(position * div_term)
+        if d_model % 2 == 0:
+            pe[:, 0, 1::2] = torch.cos(position * div_term)
+        else:
+            pe[:, 0, 1::2] = torch.cos(position * div_term[:-1])
+        pe = pe.to(device=device)
+        self.register_buffer('pe', pe)
+
+    def forward(self, x: torch.Tensor) -> torch.Tensor:
+        """
+        Arguments:
+        ----------
+        x: Tensor, shape ``[seq_len, batch_size, embedding_dim]``
+        """
+        x = x + self.pe[:x.size(0)]
+        return self.dropout(x)
+
